requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.3
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-countries==7.4.2
django-model-utils==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.3
pytz==2022.6
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.13
